requirements.txt
streamlit
tensorflow==2.15.0
numpy<2.0.0
Pillow